+++ parking/Controllers/ParkingSpaceController.cs
@@ -6,11 +6,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using parking.DTO;
 using parking.Models;
 
 namespace parking.Controllers
 {
-    internal class ParkingSpaceController
+    internal class ParkingSpaceController: DataBaseController
     {
         private PARKING_SPACE parkingSpace;
         private Helpers.Helpers h;
@@ -20,102 +21,111 @@ namespace parking.Controllers
             h = new Helpers.Helpers();
         }
 
-        public IEnumerable<dynamic> getParkingSpaces(string searchFilter = "")
+        public IEnumerable<ParkingSpaceDTO> getParkingSpaces(string searchFilter = "", bool isDel = false)
         {
-            IEnumerable<dynamic> parkingSpaces = new List<PARKING_SPACE>();
-            try
-            {
-                using (PARKINGEntities db = new PARKINGEntities())
-                {
-                    var query = from ps in db.PARKING_SPACE
-                                join pf in db.PARKING_FEE on ps.PARKING_FEE_CODE equals pf.PARKING_FEE_CODE 
-                                join pt in db.PARKING_TYPES on pf.PARKING_TYPE_CODE equals pt.PARKING_TYPE_CODE
-                                select new
-                                {
-                                    ps.PARKING_SPACE_CODE,
-                                    ps.PARKING_SPACE_NUMBER,
-                                    PARKING_STATE = ps.STATE,
-                                    ps.INSERTED_AT,
-                                    PARKING_TYPE_DESCRIPTION = pt.DESCRIPTION_PARKING_TYPE,
-                                    IS_DEL = ps.DEL
-                                };
-                    parkingSpaces = string.IsNullOrEmpty(searchFilter) ?
-                        query.ToList().Where(ps => ps.IS_DEL == false) :
-                        query.ToList().Where(ps => (ps.PARKING_SPACE_NUMBER.ToString().Contains(searchFilter) || ps.PARKING_TYPE_DESCRIPTION.Contains(searchFilter)) && ps.IS_DEL == false);
-                }
-
-            }
-            catch (Exception ex)
-            {
-                h.MsgError(ex.ToString());
-            }
-
-            return parkingSpaces;
-        }
-        public IEnumerable<dynamic> getParkingSpacesByParkingType(string parkingType)
-        {
-            IEnumerable<dynamic> parkingSpaces = new List<PARKING_SPACE>();
+            IEnumerable<ParkingSpaceDTO> parkingSpaces = new List<ParkingSpaceDTO>();
             try
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
                     var query = from ps in db.PARKING_SPACE
                                 join pf in db.PARKING_FEE on ps.PARKING_FEE_CODE equals pf.PARKING_FEE_CODE
-                                join pt in db.PARKING_TYPES on pf.PARKING_TYPE_CODE equals pt.PARKING_TYPE_CODE where pt.PARKING_TYPE_CODE == parkingType where ps.DEL == false 
-                                select new
+                                join pt in db.PARKING_TYPES on pf.PARKING_TYPE_CODE equals pt.PARKING_TYPE_CODE
+                                where string.IsNullOrEmpty(searchFilter)
+                                    ? ps.DEL == isDel
+                                    : (ps.PARKING_SPACE_NUMBER.ToString().Contains(searchFilter) ||
+                                       pt.DESCRIPTION_PARKING_TYPE.Contains(searchFilter)) &&
+                                       ps.DEL == isDel
+                                select new ParkingSpaceDTO
                                 {
-                                    ps.PARKING_SPACE_CODE,
-                                    ps.PARKING_SPACE_NUMBER,
-                                    ps.STATE,
+                                    PARKING_SPACE_CODE = ps.PARKING_SPACE_CODE,
+                                    PARKING_SPACE_NUMBER = ps.PARKING_SPACE_NUMBER,
+                                    STATE = ps.STATE,
+                                    INSERTED_AT = ps.INSERTED_AT,
+                                    PARKING_TYPE_DESCRIPTION = pt.DESCRIPTION_PARKING_TYPE,
                                     IS_DEL = ps.DEL
                                 };
+
                     parkingSpaces = query.ToList();
                 }
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
 
             return parkingSpaces;
         }
 
 
-        public dynamic getInfoParkingSpace(string id)
+        public IEnumerable<ParkingSpaceDTO> getParkingSpacesByParkingType(string parkingType)
         {
-            dynamic psp = new PARKING_SPACE();
+            IEnumerable<ParkingSpaceDTO> parkingSpaces = new List<ParkingSpaceDTO>();
             try
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
-                    var query= from ps in db.PARKING_SPACE join
-                               pf in db.PARKING_FEE on ps.PARKING_FEE_CODE equals pf.PARKING_FEE_CODE
-                                 join pt in db.PARKING_TYPES on pf.PARKING_TYPE_CODE equals pt.PARKING_TYPE_CODE
-                                 where ps.PARKING_SPACE_CODE == id
-                                 select new
-                                 {
-                                      ps.PARKING_SPACE_CODE,
-                                      ps.PARKING_SPACE_NUMBER,
-                                      ps.PARKING_FEE_CODE,
-                                      pf.PRICE_FOR_HOUR,
-                                      ps.STATE,
-                                      ps.INSERTED_AT,
-                                      pt.DESCRIPTION_PARKING_TYPE,
-                                      pt.PARKING_TYPE_CODE,
-                                 };
+                    var query = from ps in db.PARKING_SPACE
+                                join pf in db.PARKING_FEE on ps.PARKING_FEE_CODE equals pf.PARKING_FEE_CODE
+                                join pt in db.PARKING_TYPES on pf.PARKING_TYPE_CODE equals pt.PARKING_TYPE_CODE
+                                where pt.PARKING_TYPE_CODE == parkingType && ps.DEL == false
+                                select new ParkingSpaceDTO
+                                {
+                                    PARKING_SPACE_CODE = ps.PARKING_SPACE_CODE,
+                                    PARKING_SPACE_NUMBER = ps.PARKING_SPACE_NUMBER,
+                                    STATE = ps.STATE,
+                                    IS_DEL = ps.DEL
+                                };
 
-                    psp = query.FirstOrDefault();
+                    parkingSpaces = query.ToList();
                 }
-
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
+            }
+
+            return parkingSpaces;
+        }
+
+
+
+        public ParkingSpaceDTO getInfoParkingSpace(string id)
+        {
+            ParkingSpaceDTO psp = null;
+            try
+            {
+                using (PARKINGEntities db = new PARKINGEntities())
+                {
+                    var query = from ps in db.PARKING_SPACE
+                                join pf in db.PARKING_FEE on ps.PARKING_FEE_CODE equals pf.PARKING_FEE_CODE
+                                join pt in db.PARKING_TYPES on pf.PARKING_TYPE_CODE equals pt.PARKING_TYPE_CODE
+                                where ps.PARKING_SPACE_CODE == id
+                                select new ParkingSpaceDTO
+                                {
+                                    PARKING_SPACE_CODE = ps.PARKING_SPACE_CODE,
+                                    PARKING_SPACE_NUMBER = ps.PARKING_SPACE_NUMBER,
+                                    PARKING_FEE_CODE = ps.PARKING_FEE_CODE,
+                                    PRICE_FOR_HOUR = pf.PRICE_FOR_HOUR,
+                                    STATE = ps.STATE,
+                                    INSERTED_AT = ps.INSERTED_AT,
+                                    PARKING_TYPE_DESCRIPTION = pt.DESCRIPTION_PARKING_TYPE,
+                                    PARKING_TYPE_CODE = pt.PARKING_TYPE_CODE,
+                                    IS_DEL = ps.DEL
+                                };
+
+                    psp = query.FirstOrDefault();
+                }
+            }
+            catch (Exception ex)
+            {
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
 
             return psp;
         }
-        
+
+
         public PARKING_SPACE getParkingSpace(string id)
         {
             PARKING_SPACE ps= new PARKING_SPACE();
@@ -129,7 +139,7 @@ namespace parking.Controllers
             }
             catch(Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
 
             return ps;
@@ -148,7 +158,7 @@ namespace parking.Controllers
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
             return result;
 
@@ -167,7 +177,7 @@ namespace parking.Controllers
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
 
             return result;
@@ -181,6 +191,13 @@ namespace parking.Controllers
                 using(PARKINGEntities db= new PARKINGEntities())
                 {
                     PARKING_SPACE ps = db.PARKING_SPACE.Find(id);
+
+                    if (HasReferences(db,db.CHECK_IN, e => e.PARKING_SPACE_CODE==ps.PARKING_SPACE_CODE))
+                    {
+                        h.MsgError(Helpers.App.Msg0019);
+                        return 0;
+
+                    }
                     db.PARKING_SPACE.Attach(ps);
                     db.PARKING_SPACE.Remove(ps);
                     result= db.SaveChanges();
@@ -188,7 +205,7 @@ namespace parking.Controllers
 
             }catch(Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
 
             return result;
